Add option to unsort maps
